+++ frappe/public/js/frappe/form/grid.js
@@ -3,146 +3,7 @@
 
 import GridRow from "./grid_row";
 import GridPagination from "./grid_pagination";
-
-// one row of labels, same as the Data Import doctype's own template (exporter.py add_header)
-const BULK_EDIT_CSV_HEADER_ROWS = 1;
-const BULK_EDIT_MAX_ROWS = 5000;
-const BULK_EDIT_FILE_TYPES = [".csv", ".xlsx", ".xls"];
-const BULK_EDIT_ID_FIELDNAME = "name";
-// same labels as Data Import, so the strings are already translated
-const BULK_EDIT_INSERT = "Insert New Records";
-const BULK_EDIT_UPDATE = "Update Existing Records";
-const BULK_EDIT_UPSERT = "Insert or Update Records";
-const BULK_EDIT_IMPORT_TYPES = [BULK_EDIT_INSERT, BULK_EDIT_UPDATE, BULK_EDIT_UPSERT];
-const BULK_EDIT_DONT_IMPORT = "Don't Import";
-const BULK_EDIT_BLANK_TEMPLATE = "blank_template";
-const BULK_EDIT_ALL_RECORDS = "all";
-const BULK_EDIT_5_RECORDS = "5_records";
-// same pattern as DURATION_PATTERN in importer.py (frappe/core/doctype/data_import/importer.py)
-const BULK_EDIT_DURATION_PATTERN = /^(?:(\d+d)?((^|\s)\d+h)?((^|\s)\d+m)?((^|\s)\d+s)?)$/;
-// a Duration is stored as a number of seconds, which is what the template
-// exports, so a file coming back round-trip carries digits rather than "1h 30m"
-const BULK_EDIT_SECONDS_PATTERN = /^\d+$/;
-// the words importer.py's Row.parse_value accepts for a Check field, alongside 0/1
-const BULK_EDIT_CHECK_TRUE = ["t", "true", "y", "yes"];
-const BULK_EDIT_CHECK_FALSE = ["f", "false", "n", "no"];
-const BULK_EDIT_NUMERIC_FIELDTYPES = ["Int", "Float", "Currency", "Percent"];
-// Fieldtypes whose control refuses a value it cannot read: it parses to null or
-// "Invalid date" and then renders blank. A flagged cell holds exactly such a
-// value, so these controls are told to keep the text as the file wrote it.
-// Fieldtypes whose control cannot show a value it cannot read, and paints over
-// the cell rather than leave it alone: a datepicker rewrites the input from its
-// own state ("NaN:NaN:NaN" out of ControlTime.set_input, an invalid date out of
-// ControlDate.set_formatted_input), a duration picker resolves every box to NaN
-// (seconds_to_duration of a word), and a checkbox can draw no word at all. A
-// flagged cell of one of these keeps the file's own text on show and builds its
-// control on the click that goes to resolve it.
-const BULK_EDIT_DEFERRED_FIELDTYPES = ["Date", "Datetime", "Time", "Duration", "Check"];
-const BULK_EDIT_RAW_TEXT_FIELDTYPES = [
-	...BULK_EDIT_NUMERIC_FIELDTYPES,
-	"Date",
-	"Datetime",
-	"Time",
-	"Duration",
-];
-// every step shares one size, so switching tabs never resizes the modal.
-// the modal is sized against the window rather than in pixels: it takes 90%
-// of the height, less the header and footer the body sits between, so the
-// margin above and below stays even on any screen. the width comes with it,
-// from .bulk-edit-dialog in grid.scss
-const BULK_EDIT_DIALOG_SIZE = "extra-large";
-const BULK_EDIT_DIALOG_HEIGHT = "calc(90vh - 104px)";
-const BULK_EDIT_PREVIEW_ROWS = 10;
-// spreadsheet cells come back in system format, csv cells in the user's date format
-const SYSTEM_DATE_PATTERN = /^\d{4}-\d{2}-\d{2}/;
-
-// Every fieldtype whose stored value is not the string the file carries. These
-// mirror importer.py Row.parse_value — a value that passed validation still has
-// to be converted, or it reaches the doc as text and is silently wrong.
-const BULK_EDIT_VALUE_FORMATTERS = {
-	Date: (val) => {
-		if (!val) return val;
-		return SYSTEM_DATE_PATTERN.test(val) ? val.slice(0, 10) : frappe.datetime.user_to_str(val);
-	},
-	Datetime: (val) => (val ? frappe.datetime.user_to_str(val) : val),
-	Time: (val) => bulk_edit_to_system_time(val),
-	Int: (val) => cint(val),
-	Check: (val) => {
-		const word = cstr(val).trim().toLowerCase();
-		if (BULK_EDIT_CHECK_TRUE.includes(word)) return 1;
-		if (BULK_EDIT_CHECK_FALSE.includes(word)) return 0;
-		return cint(val);
-	},
-	Float: (val) => flt(val),
-	Currency: (val) => flt(val),
-	Percent: (val) => flt(val),
-	Duration: (val) => bulk_edit_to_seconds(val),
-};
-
-/** "1d 2h 30m" to seconds, via the same helper the Duration control uses. */
-function bulk_edit_to_seconds(value) {
-	const text = cstr(value).trim();
-	if (!text) return 0;
-	// already seconds — the template exports the stored value, not a duration string
-	if (BULK_EDIT_SECONDS_PATTERN.test(text)) return cint(text);
-
-	const part = (unit) => cint((text.match(new RegExp(`(\\d+)${unit}`)) || [])[1]);
-	return frappe.utils.duration_to_seconds(part("d"), part("h"), part("m"), part("s"));
-}
-
-/**
- * What a freshly mounted cell control starts on. The file's own value goes in
- * unchanged wherever the control can hold it; where it cannot, the control is
- * seeded clean and the cell's own text is what keeps the faulty value on show.
- */
-function bulk_edit_seed_value(fieldtype, value) {
-	// ControlCheck.set_input() runs the value through cint(), reading "Yes" as 0
-	if (fieldtype === "Check") return BULK_EDIT_VALUE_FORMATTERS.Check(value);
-	// the rest of the deferred set is only ever mounted over a flagged cell, so the
-	// value here is one the widget cannot read; starting clean opens the picker on
-	// a real time rather than on nonsense, and the cell's own text is what kept the
-	// faulty value on show until this click
-	if (BULK_EDIT_DEFERRED_FIELDTYPES.includes(fieldtype)) return "";
-	return value;
-}
-
-/** A time in the user's format to the "HH:mm:ss" the doc stores. */
-function bulk_edit_to_system_time(value) {
-	if (!value) return value;
-	const parsed = moment(value, BULK_EDIT_TIME_FORMATS(), true);
-	return parsed.isValid() ? parsed.format(frappe.defaultTimeFormat) : value;
-}
-
-/** The two time formats a file may carry: the user's own, and the stored one. */
-const BULK_EDIT_TIME_FORMATS = () => [
-	frappe.datetime.get_user_time_fmt(),
-	frappe.defaultTimeFormat,
-];
-
-/**
- * Whether the whole value reads as a number. Written against flt()'s own steps —
- * the same currency strip and the same strip_number_groups — so the locale's
- * group and decimal separators are read here exactly as they will be when the
- * value is applied, which a regex of our own would not manage.
- *
- * The last step is deliberately stricter than flt: parseFloat stops at the first
- * character it cannot read, so flt turns "12abc" into 12 and "1.2.3" into 1.2.
- * Number() rejects both, which is the point — a value that only half parses is
- * the kind that imports quietly wrong.
- */
-function bulk_edit_is_number(value) {
-	let text = cstr(value).trim();
-	if (!text) return false;
-
-	// flt drops a currency symbol when a space separates it from the figure
-	if (text.includes(" ")) {
-		const parts = text.split(" ");
-		if (isNaN(parseFloat(parts[0]))) text = parts.slice(parts.length - 1).join(" ");
-	}
-
-	text = strip_number_groups(text);
-	return text !== "" && !isNaN(Number(text));
-}
+import BulkEdit from "./bulk_edit";
 
 // Static pixel column widths; legacy map migrates old 1-12 `columns`/`colsize`.
 export const GRID_MIN_COLUMN_WIDTH = 60;
@@ -1792,1408 +1653,9 @@ export default class Grid {
 			.find(".grid-upload")
 			.removeClass("hidden")
 			.on("click", () => {
-				this.show_bulk_edit_wizard();
+				new BulkEdit(this).show();
 				return false;
 			});
-	}
-
-	get_bulk_edit_title() {
-		return this.df.label || frappe.model.unscrub(this.df.fieldname);
-	}
-
-	/**
-	 * Value fields of the child doctype, with ID first so rows can be matched on
-	 * it. Read-only fields are left out: the document's own code writes them on
-	 * save (amount = rate * qty, and so on), so a column for one is a column
-	 * whose values never survive. This is the single gate — it feeds the download
-	 * template and the mapping picker alike, so nothing downstream has to test
-	 * for read_only again. Its server-side twin is in bulk_edit.py.
-	 */
-	get_bulk_edit_docfields() {
-		return [
-			{ fieldname: BULK_EDIT_ID_FIELDNAME, label: __("ID"), fieldtype: "Data" },
-			...frappe
-				.get_meta(this.df.options)
-				.fields.filter((df) => frappe.model.is_value_type(df.fieldtype) && !df.read_only),
-		];
-	}
-
-	/**
-	 * One modal, one size, three steps: Setup, Upload, Preview. A step only
-	 * opens once the one before it has produced what it needs, and stays open
-	 * afterwards so earlier choices can be revisited without starting again.
-	 * Mapping, the per-cell fixes and skipping all happen in Preview, so
-	 * nothing has to be corrected a step away from where it's shown. Nothing
-	 * opens on top of the modal, and every step's actions sit in the same
-	 * place in the footer.
-	 */
-	show_bulk_edit_wizard() {
-		// a read only grid cannot take rows back, so only the template step applies
-		const can_import = this.is_editable();
-
-		const state = {
-			import_type: BULK_EDIT_INSERT,
-			headers: [],
-			rows: [],
-			row_numbers: [],
-			column_map: {},
-			warnings: [],
-			// row numbers left out of the import by their own checkbox in the preview
-			skipped_rows: new Set(),
-		};
-
-		const panels = {
-			setup: $('<div class="bulk-edit-panel"></div>'),
-			upload: $('<div class="bulk-edit-panel"></div>'),
-			// mapping and inline fixes both live here — one step, not a
-			// separate Preview/Fix Issues pair
-			preview: $('<div class="bulk-edit-panel"></div>'),
-		};
-
-		const dialog = new frappe.ui.Dialog({
-			title: __("Upload {0}", [this.get_bulk_edit_title()]),
-			size: BULK_EDIT_DIALOG_SIZE,
-			centered: true,
-		});
-		// the width is a share of the window too, so the modal keeps the same
-		// proportions against the page that the height does
-		$(dialog.wrapper).addClass("bulk-edit-dialog");
-		// the body owns the height and each panel fills what is left under the tab
-		// bar, so a tall panel scrolls inside itself instead of growing the modal
-		dialog.modal_body.css({ height: BULK_EDIT_DIALOG_HEIGHT, "overflow-y": "hidden" });
-		dialog.$body.css({ height: "100%", display: "flex", "flex-direction": "column" });
-
-		const tab_defs = can_import
-			? [
-					{ label: __("Setup"), content: () => panels.setup[0] },
-					{ label: __("Upload"), content: () => build_uploader(), disabled: true },
-					{ label: __("Preview"), content: () => panels.preview[0], disabled: true },
-				]
-			: [{ label: __("Setup"), content: () => panels.setup[0] }];
-
-		const tabs = new frappe.ui.Tabs({
-			tabs: tab_defs,
-			on_change: (index) => {
-				stepper.set_current(index);
-				sync_uploaded_file();
-				set_footer();
-			},
-		});
-		tabs.$el.addClass("bulk-edit-tabs");
-		// Tabs' own bar still exists (it drives the panels) but the Stepper
-		// below replaces it visually — same component the Data Import wizard
-		// uses for its step row, so the two dialogs read identically.
-		tabs.$el.find(".es-tabs__list").hide();
-
-		const stepper = new frappe.ui.Stepper({
-			steps: tab_defs.map((tab) => ({ label: tab.label })),
-			is_locked: (index) => tab_defs[index].disabled,
-			on_step_click: (index) => tabs.set_active(index),
-		});
-
-		// tabs.set_disabled() is called from several places below to lock/unlock
-		// steps as the flow progresses; wrapping it here means the stepper's
-		// lock icons stay in sync everywhere, without touching each call site.
-		const set_disabled = tabs.set_disabled.bind(tabs);
-		tabs.set_disabled = (index, disabled) => {
-			set_disabled(index, disabled);
-			stepper.refresh();
-		};
-
-		// The step row reads on the modal's grey body; the panel it governs and
-		// the dialog's own footer actions live together in one white card
-		// underneath it — same shape as the Data Import wizard's dialog. Moving
-		// dialog.footer's actual node (not rebuilding it) means set_footer()'s
-		// primary/secondary action calls below need no changes.
-		const $card = $('<div class="bulk-edit-card"></div>');
-		dialog.$body.append(stepper.$el, $card);
-		$card.append(tabs.$el, dialog.footer);
-
-		tabs.$el.css({
-			flex: "1 1 auto",
-			"min-height": 0,
-			display: "flex",
-			"flex-direction": "column",
-		});
-		tabs.$el
-			.find(".es-tabs__panel")
-			.css({ flex: "1 1 auto", "min-height": 0, "overflow-y": "auto" });
-
-		const TAB_SETUP = 0;
-		const TAB_UPLOAD = 1;
-		const TAB_PREVIEW = 2;
-
-		// ---------------------------------------------------------- step 1
-
-		const setup_form = new frappe.ui.FieldGroup({
-			body: panels.setup[0],
-			no_submit_on_enter: true,
-			fields: [
-				{
-					fieldtype: "Select",
-					fieldname: "import_type",
-					label: __("Import Type"),
-					options: BULK_EDIT_IMPORT_TYPES.map((value) => ({ label: __(value), value })),
-					default: BULK_EDIT_INSERT,
-					reqd: 1,
-					// the export that suits the import, still free to change afterwards
-					change: () => {
-						const value = setup_form.get_value("import_type");
-						state.import_type = value;
-						setup_form.set_value(
-							"export_records",
-							value === BULK_EDIT_INSERT
-								? BULK_EDIT_BLANK_TEMPLATE
-								: BULK_EDIT_ALL_RECORDS,
-						);
-						tabs.set_disabled(TAB_UPLOAD, !value);
-					},
-				},
-				{ fieldtype: "Section Break" },
-				{
-					fieldtype: "Select",
-					fieldname: "file_type",
-					label: __("File Type"),
-					options: ["Excel", "CSV"],
-					default: "Excel",
-					reqd: 1,
-				},
-				{ fieldtype: "Column Break" },
-				{
-					fieldtype: "Select",
-					fieldname: "export_records",
-					label: __("Export Type"),
-					options: [
-						{ label: __("Blank Template"), value: BULK_EDIT_BLANK_TEMPLATE },
-						{ label: __("All Records"), value: BULK_EDIT_ALL_RECORDS },
-						{ label: __("5 Records"), value: BULK_EDIT_5_RECORDS },
-					],
-					default: BULK_EDIT_BLANK_TEMPLATE,
-					description: __("{0} rows in this table", [
-						(this.frm.doc[this.df.fieldname] || []).length,
-					]),
-				},
-				{ fieldtype: "Section Break" },
-				{
-					fieldtype: "MultiCheck",
-					fieldname: "fields",
-					label: __("Fields"),
-					columns: 2,
-					select_all: true,
-					select_mandatory: true,
-					sort_options: false,
-					options: this.get_bulk_edit_docfields().map((df) => {
-						// ID and mandatory fields only; row-matching needs ID and the rest is opt-in
-						const mandatory = df.fieldname === BULK_EDIT_ID_FIELDNAME || !!df.reqd;
-						return {
-							label: __(df.label || df.fieldname, null, df.parent),
-							value: df.fieldname,
-							checked: mandatory ? 1 : 0,
-							danger: mandatory,
-						};
-					}),
-					on_change: () => set_footer(),
-				},
-			],
-		});
-		setup_form.make();
-
-		// ---------------------------------------------------------- step 2
-
-		// mounted inline rather than in its own dialog, so nothing stacks
-		let file_uploader = null;
-		const uploaded_file_count = () => file_uploader?.uploader?.files?.length || 0;
-
-		// File upload / Google Sheet as tabs, same source picker as the Data
-		// Import wizard's Config step (frappe.ui.Tabs, same labels and icons).
-		const build_uploader = () => {
-			panels.upload.css({ height: "100%", display: "flex", "flex-direction": "column" });
-
-			const $file_pane = $('<div class="bulk-edit-upload-pane bulk-edit-file-pane"></div>');
-			const $sheet_pane = $('<div class="bulk-edit-upload-pane"></div>');
-
-			const upload_tabs = new frappe.ui.Tabs({
-				css_class: "bulk-edit-upload-tabs",
-				tabs: [
-					{ label: __("File upload"), icon: "upload", content: $file_pane[0] },
-					{ label: __("Google Sheet"), icon: "link", content: $sheet_pane[0] },
-				],
-			});
-			panels.upload.append(upload_tabs.$el);
-			upload_tabs.$el.css({
-				flex: "1 1 auto",
-				"min-height": 0,
-				display: "flex",
-				"flex-direction": "column",
-			});
-			// the bar keeps its own height regardless of what the panel below it
-			// does — without this it's free to shrink (flex's default) and the
-			// growing panel squeezes the tab labels
-			upload_tabs.$el.find(".es-tabs__list").css({ flex: "0 0 auto" });
-			// sizing only, not display — .es-tabs__panel's own display is how the
-			// tab component hides the inactive pane ([data-state="inactive"]), and
-			// an inline display here would win over that and show both at once.
-			// overflow-y matters too: the file pane centres its content, and
-			// without a scroll container of its own, anything taller than the
-			// available space bleeds out both ways — up, over the tab bar, included
-			upload_tabs.$el.find(".es-tabs__panel").css({
-				flex: "1 1 auto",
-				"min-height": 0,
-				"overflow-y": "auto",
-			});
-			// centring (empty) vs. top-aligned (.has-file, toggled below) is in
-			// grid.scss; the Google Sheet pane is a plain field, so it's always top
-			$file_pane.css({ height: "100%" });
-			$sheet_pane.css({ height: "100%" });
-
-			file_uploader = new frappe.ui.FileUploader({
-				wrapper: $file_pane,
-				as_dataurl: true,
-				allow_multiple: false,
-				// none of these fit a CSV/Excel-only upload — Link duplicates the
-				// Google Sheet tab, and Camera/Google Drive don't produce spreadsheets.
-				// Library (the internal file browser) stays on, same as the PR.
-				allow_web_link: false,
-				allow_take_photo: false,
-				allow_google_drive: false,
-				restrictions: { allowed_file_types: BULK_EDIT_FILE_TYPES },
-				on_success: (file) => this.read_bulk_edit_file(file, on_file),
-			});
-			// keep it at its natural height so the centring above has room to work
-			$file_pane.children(".file-uploader").css({ flex: "0 0 auto" });
-			// dropping a file or clearing one changes what the later tabs describe,
-			// and whether the pane still reads as an empty drop target
-			$file_pane.on("click change drop", () =>
-				setTimeout(() => {
-					$file_pane.toggleClass("has-file", Boolean(uploaded_file_count()));
-					sync_uploaded_file();
-					set_footer();
-				}, 0),
-			);
-
-			// No separate Import button — same as the PR: entering a URL and
-			// leaving the field (change fires on blur/Enter) is the trigger.
-			const sheet_form = new frappe.ui.FieldGroup({
-				body: $sheet_pane[0],
-				no_submit_on_enter: true,
-				fields: [
-					{
-						// same label/description as the Data Import doctype's own
-						// google_sheets_url field, for the same reason it uses them
-						fieldtype: "Data",
-						fieldname: "google_sheets_url",
-						label: __("Import from Google Sheets"),
-						description: __("Must be a publicly accessible Google Sheets URL"),
-						change: () => {
-							const url = sheet_form.get_value("google_sheets_url");
-							if (url) this.read_bulk_edit_google_sheet(url, on_file);
-						},
-					},
-				],
-			});
-			sheet_form.make();
-
-			return panels.upload[0];
-		};
-
-		/** Clearing the file leaves nothing for preview or mapping to describe. */
-		const sync_uploaded_file = () => {
-			if (!file_uploader || uploaded_file_count() || !state.rows.length) return;
-			state.headers = [];
-			state.rows = [];
-			state.row_numbers = [];
-			state.column_map = {};
-			tabs.set_disabled(TAB_PREVIEW, true);
-		};
-
-		// ---------------------------------------------------------- step 3
-
-		let preview_form = null;
-		let mapping_controls = [];
-		// set_value fires the control's change hook, so the initial pass would
-		// redraw once per column before the last one exists. One redraw at the end
-		// of the build covers them all.
-		let building_preview = false;
-		// discards a stale link-check response if a newer mapping change started one first
-		let preview_request_id = 0;
-		// live edit controls mounted over an invalid Link/Select cell, keyed "row:col"
-		let cell_controls = {};
-		/**
-		 * Drop a control's picker along with the control. air-datepicker mounts into a
-		 * container on the body, so one left behind outlives the cell it belonged to —
-		 * still on screen, with nothing left that could close it.
-		 */
-		const discard_cell_control = (control) => {
-			control?.hide_picker?.();
-			control?.datepicker?.destroy?.();
-		};
-		const discard_cell_controls = () =>
-			Object.values(cell_controls).forEach(discard_cell_control);
-
-		/**
-		 * One listener for every cell picker rather than one each, so a control whose
-		 * cell has since been redrawn is still reached — its picker lives in the body,
-		 * not in the cell.
-		 */
-		const on_dialog_mousedown = (event) => {
-			Object.values(cell_controls).forEach((control) => {
-				const picker = picker_of(control);
-				// no picker, or the click is someone using it
-				if (!picker || picker.contains(event.target)) return;
-
-				if (control.$wrapper.closest("td").get(0)?.contains(event.target)) {
-					// the cell toggles, since the control's own handler only ever opens
-					if (!picker_is_open(control)) {
-						open_picker(control);
-						return;
-					}
-					close_picker(control);
-					// Closing alone is not enough: both pickers open on the input's focus
-					// (air-datepicker's showEvent, and ControlDuration's own focus handler),
-					// and air-datepicker's hide() blurs the input on its way out — so the
-					// focus this mousedown is about to deliver would open it straight back
-					// up. Taking the default away leaves the input unfocused and the picker
-					// shut; the next click on the cell opens it again through the branch
-					// above.
-					event.preventDefault();
-					return;
-				}
-				if (picker_is_open(control)) close_picker(control);
-			});
-		};
-		// capture, so it runs before the dialog or the table can swallow the click
-		document.addEventListener("mousedown", on_dialog_mousedown, true);
-		dialog.$wrapper.on("hidden.bs.modal", () => {
-			discard_cell_controls();
-			document.removeEventListener("mousedown", on_dialog_mousedown, true);
-		});
-
-		/** "Don't Import", then every field a column can land in. */
-		const mapping_options = () => [
-			{ label: __("Don't Import"), value: BULK_EDIT_DONT_IMPORT },
-			...this.get_bulk_edit_docfields().map((df) => ({
-				label: __(df.label || df.fieldname, null, df.parent),
-				value: df.fieldname,
-				// the file speaks fieldnames, so the picker shows both names
-				description: df.fieldname,
-			})),
-		];
-
-		/**
-		 * The preview is the mapping step: every column of the file gets a picker
-		 * in the header cell above its own values, so a column and the field it
-		 * lands in are read together. Built once per file — a mapping change only
-		 * re-reads the pickers, which keeps them (and any open dropdown) alive.
-		 */
-		const build_preview = () => {
-			discard_cell_controls();
-			panels.preview.empty();
-			cell_controls = {};
-			state.skipped_rows = new Set();
-			preview_form = new frappe.ui.FieldGroup({
-				body: panels.preview[0],
-				no_submit_on_enter: true,
-				fields: [{ fieldtype: "HTML", fieldname: "table" }],
-			});
-			preview_form.make();
-
-			const $table = preview_form.get_field("table").$wrapper;
-			$table.html(
-				this.get_bulk_edit_preview_html(state.headers, state.rows, state.row_numbers),
-			);
-			const options = mapping_options();
-			building_preview = true;
-			mapping_controls = state.headers.map((header, i) => {
-				const control = frappe.ui.form.make_control({
-					df: {
-						fieldtype: "Autocomplete",
-						fieldname: `map_${i}`,
-						placeholder: header || __("Column {0}", [i + 1]),
-						max_items: Infinity,
-						options,
-						change: () => refresh_preview(),
-					},
-					parent: $table.find(`.bulk-edit-mapping-row td[data-col="${i}"]`).get(0),
-					render_input: true,
-					only_input: true,
-				});
-				// this list is every field of the child doctype, so it is long and
-				// has to scroll; unpinned it is clipped by the table's own overflow
-				pin_dropdown(control);
-				control.set_value(state.column_map[i] || BULK_EDIT_DONT_IMPORT);
-				return control;
-			});
-			building_preview = false;
-
-			refresh_preview();
-		};
-
-		/**
-		 * What's wrong with the cell being edited, in the one shared line under
-		 * the "Map each column..." text — not inside the cell, since only_input
-		 * mode (needed for the borderless in-table look) suppresses
-		 * set_description() outright anyway. title as well as text: the gap is a
-		 * single clipped line, so a long message stays readable on hover.
-		 * Reads _warning live, so it says what is true now.
-		 */
-		const show_cell_message = (control) => {
-			const message = control?._warning?.message || "";
-			preview_form
-				.get_field("table")
-				.$wrapper.find(".bulk-edit-preview-message")
-				.text(message)
-				.attr("title", message);
-		};
-
-		/**
-		 * Pin an open dropdown to where its input is on screen. The list sits in a
-		 * container that scrolls and clips, and freeing it by switching that
-		 * container's overflow moves the whole grid (see grid.scss). Fixed takes it
-		 * out of every ancestor's flow and clipping instead, so nothing reflows.
-		 * Re-pinned on each open, since the row may have moved; closed on scroll,
-		 * since a fixed element doesn't travel with the container it came from.
-		 */
-		const pin_dropdown = (control) => {
-			const list = () => control.$input.closest(".awesomplete").children("ul").get(0);
-			// The list is closed when the page moves under it, since a fixed element
-			// doesn't travel with the container it came from. Its own scrolling is
-			// not that: the list is taller than its max-height and scrolls inside
-			// itself, and closing on that would leave the last options reachable
-			// only by typing.
-			const close = (event) => {
-				if (event && list()?.contains(event.target)) return;
-				control.awesomplete?.close();
-			};
-			control.$input.on("awesomplete-open", () => {
-				const rect = control.$input[0].getBoundingClientRect();
-				$(list()).css({
-					position: "fixed",
-					top: `${rect.bottom}px`,
-					left: `${rect.left}px`,
-					width: `${rect.width}px`,
-					"min-width": 0,
-					// above the dialog and its sticky footer, which a fixed element
-					// is no longer stacked against by nesting alone
-					"z-index": 1050,
-				});
-				// capture: scroll doesn't bubble, and what scrolls here is an ancestor
-				document.addEventListener("scroll", close, true);
-			});
-			control.$input.on("awesomplete-close", () =>
-				document.removeEventListener("scroll", close, true),
-			);
-		};
-
-		/**
-		 * Date, Datetime and Time carry an air-datepicker, Duration its own box; both
-		 * mount in the body, open on focus, and leave closing to the input's blur —
-		 * which never comes while that input keeps focus. So the picker hung over the
-		 * table when the same cell was clicked again, or the dialog around it was.
-		 */
-		const picker_of = (control) =>
-			control?.$picker?.get(0) || control?.datepicker?.$datepicker?.get(0) || null;
-		const picker_is_open = (control) =>
-			control.$picker
-				? control.$picker.is(":visible")
-				: Boolean(control.datepicker?.visible);
-		const open_picker = (control) =>
-			control.$picker ? control.show_picker() : control.datepicker?.show();
-		const close_picker = (control) =>
-			control.$picker ? control.hide_picker() : control.datepicker?.hide();
-
-		/**
-		 * A control of the field's own type, mounted over a cell so a value the
-		 * file got wrong can be seen and corrected in place. Every fieldtype's own
-		 * "reject what doesn't fit" behaviour has to be turned off first, or it
-		 * blanks the cell the moment the value is set.
-		 */
-		const make_cell_control = (cell, r, col, warning, fieldname) => {
-			const original = state.rows[r][col];
-			const df = { ...warning.field };
-
-			// A hidden field renders no control at all (base_control.js get_status
-			// returns "None"), which would leave the cell looking empty and dead.
-			// Hidden describes how the field behaves on a form; the value here is the
-			// file's, still being corrected on its way in, and the import writes it
-			// either way. Read-only needs no such reset — those fields are never
-			// offered for mapping (get_bulk_edit_docfields), so no cell is ever one.
-			df.hidden = 0;
-			df.hidden_due_to_dependency = 0;
-
-			if (df.fieldtype === "Select") {
-				// a <select> can't display a value with no matching <option> — add
-				// the file's own value as one, so it shows instead of blank
-				const options = (df.options || "").split("\n").map((o) => o.trim());
-				if (original && !options.includes(cstr(original).trim())) {
-					df.options = [original, ...options].join("\n");
-				}
-			} else if (df.fieldtype === "Link") {
-				// ControlLink.validate() existence-checks against the server and
-				// returns empty when there's no such record — exactly our case. The
-				// value's invalidity is already reported by our own warning.
-				df.ignore_link_validation = true;
-			}
-
-			// the input fills the whole cell (grid.scss) rather than sitting inside
-			// it, so any click in the cell is a genuine click on the input — the
-			// only way a native <select> reliably opens its list across browsers is
-			// a real click, not a programmatic .focus()/.click()
-			$(cell).addClass("bulk-edit-editable-cell");
-			const control = frappe.ui.form.make_control({
-				df: {
-					...df,
-					change: () => {
-						state.rows[r][col] = control.get_value();
-						refresh_preview();
-					},
-				},
-				parent: $(cell).empty().get(0),
-				render_input: true,
-				only_input: true,
-			});
-			control._fieldname = fieldname;
-			control._warning = warning;
-			// what the dropdown cue below positions itself against
-			control.$wrapper.css("position", "relative");
-
-			if (df.fieldtype === "Link") {
-				// an invalid value has no record to fetch a title for, which would
-				// otherwise blank the input — show it as plain text, same as every
-				// other fieldtype already does
-				control.set_link_title = async (value) =>
-					control.translate_and_set_input_value(value, value);
-				// Select gets a dropdown cue for free (ControlSelect.set_icon);
-				// Link's autocomplete input doesn't. Inline position, not a
-				// stylesheet rule: .select-icon's own CSS needs a positioned
-				// ancestor this cell's markup doesn't reliably give it.
-				$(`<div class="select-icon">${frappe.utils.icon("chevrons-up-down", "sm")}</div>`)
-					.css({
-						position: "absolute",
-						top: "3px",
-						right: "12px",
-						"pointer-events": "none",
-					})
-					.appendTo(control.$wrapper);
-				pin_dropdown(control);
-			}
-
-			if (BULK_EDIT_RAW_TEXT_FIELDTYPES.includes(df.fieldtype)) {
-				// Each of these throws the value away rather than show one it cannot
-				// read: ControlFloat.parse() returns null, ControlInt.parse() turns
-				// "abc" into 0, and ControlDate/Datetime/Time return "" on "Invalid
-				// date". The cell would then come up blank on the very value it is
-				// flagged for, leaving nothing to see or correct. Hold the text as the
-				// file wrote it; converting it is BULK_EDIT_VALUE_FORMATTERS' job,
-				// once, on apply. validate() goes with them — ControlDate's calls
-				// frappe.datetime.validate() and blanks the input just as surely.
-				control.parse = (value) => value;
-				control.format_for_input = (value) => cstr(value);
-				control.validate = (value) => value;
-			}
-
-			control.set_value(bulk_edit_seed_value(df.fieldtype, original));
-
-			// click as well as focus: focus fires once and the cell stays focused,
-			// so every click after the first would otherwise do nothing. Link's own
-			// focus handler (link.js) only opens the list when the input is empty,
-			// and ours never is — it always holds the file's value — so opening has
-			// to be asked for explicitly, every time. The empty term matters:
-			// on_input() searches on whatever the input holds, which here is the
-			// value no record matches, so the list would come up empty; searching on
-			// "" lists the real options while the cell goes on showing what the file
-			// said. link.js sets filter: () => true, so awesomplete won't filter
-			// them back out.
-			control.$input?.on("focus click", () => {
-				show_cell_message(control);
-				if (df.fieldtype === "Link") control.on_input({ target: { value: "" } });
-			});
-
-			return control;
-		};
-
-		/**
-		 * A row nothing can be done about is skippable: leaving it out is what lets
-		 * the rest of the file through, since its warnings stop counting against
-		 * Apply once it is (see set_footer). Only rows that actually have something
-		 * wrong get a button — plus any already skipped, so there's a way back.
-		 */
-		const render_skip_buttons = ($table, warnings) => {
-			const rows_with_warnings = new Set(
-				warnings.filter((w) => w.row !== undefined).map((w) => cint(w.row)),
-			);
-			$table.find("tr[data-row]").each((_, tr) => {
-				const row = cint(tr.dataset.row);
-				const skipped = state.skipped_rows.has(row);
-				const $cell = $(tr).find(".bulk-edit-skip-cell").empty();
-				$(tr).toggleClass("bulk-edit-skipped-row", skipped);
-				if (!skipped && !rows_with_warnings.has(row)) return;
-
-				frappe.ui
-					.button({
-						label: skipped ? __("Restore") : __("Skip"),
-						// same options this grid's own footer buttons use (Add row,
-						// Edit, Duplicate rows): default subtle variant, size sm, and
-						// no red theme — skipping a row is an ordinary action, not a
-						// destructive one
-						size: "sm",
-						onclick: () => {
-							state.skipped_rows[skipped ? "delete" : "add"](row);
-							refresh_preview();
-						},
-					})
-					.appendTo($cell);
-			});
-		};
-
-		/**
-		 * Mark every cell a warning names — same has-error convention the doctype's
-		 * own editable grid rows use — and give each one a live control of its own
-		 * fieldtype, so the value can be fixed without leaving the preview. The
-		 * control stays for as long as the column maps to that same field.
-		 */
-		const sync_preview_errors = (warnings) => {
-			const by_cell = {};
-			warnings.forEach((w) => {
-				if (w.row !== undefined && w.col !== undefined) by_cell[`${w.row}:${w.col}`] = w;
-			});
-
-			const $table = preview_form.get_field("table").$wrapper;
-			render_skip_buttons($table, warnings);
-
-			// tr[data-row] scopes this to the data rows: the mapping row carries
-			// td[data-col] cells of its own, and they hold pickers, not values
-			$table.find("tr[data-row] td[data-col]").each((_, cell) => {
-				const row = cint(cell.closest("tr").dataset.row);
-				const col = cint(cell.dataset.col);
-				const key = `${row}:${col}`;
-				const warning = by_cell[key];
-				const fieldname = state.column_map[col];
-				const existing = cell_controls[key];
-
-				// red only while the value is actually invalid — the control
-				// underneath stays put either way
-				$(cell).toggleClass("has-error", Boolean(warning));
-
-				if (existing) {
-					// still the same target field: keep the control and just keep its
-					// warning current, so re-editing is never a one-shot thing
-					if (existing._fieldname === fieldname) {
-						existing._warning = warning;
-						return;
-					}
-					// the column got remapped — this control is for the field that
-					// used to be here, so drop it and fall through to build a fresh
-					// one for what's mapped now
-					discard_cell_control(existing);
-					delete cell_controls[key];
-					$(cell)
-						.removeClass("bulk-edit-editable-cell")
-						.empty()
-						.text(state.rows[state.row_numbers.indexOf(row)][col]);
-				}
-
-				const r = state.row_numbers.indexOf(row);
-				const mapped_df =
-					fieldname && frappe.meta.get_docfield(this.df.options, fieldname);
-				const fieldtype = mapped_df?.fieldtype;
-				$(cell).removeClass("bulk-edit-pending-cell").off("click.bulk-edit-reveal");
-
-				// A picker or a checkbox cannot show a value it cannot read — it draws
-				// its own state over the input instead, which is where "NaN:NaN:NaN"
-				// came from. So the faulty value stays on show as the cell's own text,
-				// in the error colour, and the control is built by the click that goes
-				// to resolve it: seeded clean, so the picker opens on a real time.
-				if (warning?.field && BULK_EDIT_DEFERRED_FIELDTYPES.includes(fieldtype)) {
-					$(cell)
-						.addClass("bulk-edit-pending-cell")
-						.one("click.bulk-edit-reveal", () => {
-							const control = make_cell_control(cell, r, col, warning, fieldname);
-							cell_controls[key] = control;
-							// the click that revealed it landed on the cell, not on the control
-							control.$input?.trigger("focus");
-						});
-					return;
-				}
-
-				// a Check whose value the box can represent reads as the box itself:
-				// "Yes" and "1" both mean a tick, which is what will be imported
-				if (!warning?.field && fieldtype !== "Check") return;
-
-				cell_controls[key] = make_cell_control(
-					cell,
-					r,
-					col,
-					warning || { field: mapped_df },
-					fieldname,
-				);
-			});
-
-			// the edit that prompted this recompute may well have fixed the very
-			// cell being edited, in which case the sentence saying what was wrong
-			// with it has to go too. Focus doesn't move when a value is picked, so
-			// the handler above won't fire again to clear it.
-			show_cell_message(Object.values(cell_controls).find((c) => c.$input?.is(":focus")));
-		};
-
-		/**
-		 * Re-read the pickers, then redraw everything that describes the mapping:
-		 * the column states in the preview, the counts and the warnings.
-		 */
-		const refresh_preview = async () => {
-			if (building_preview) return;
-			const request_id = ++preview_request_id;
-
-			const map = {};
-			mapping_controls.forEach((control, i) => {
-				const value = control.get_value();
-				if (value && value !== BULK_EDIT_DONT_IMPORT) map[i] = value;
-			});
-			state.column_map = map;
-
-			// a mapped column carries its values into the table, an unmapped one is
-			// along for the ride; the cells say so without a legend. The header
-			// text itself is the file's own and never changes with the mapping —
-			// which field a column lands in is the picker's job to show, one row
-			// down, and a column stays recognisable by what the file called it.
-			preview_form
-				.get_field("table")
-				.$wrapper.find("[data-col]")
-				.each((_, cell) => {
-					$(cell).attr("data-mapped", map[cint(cell.dataset.col)] ? 1 : 0);
-				});
-
-			const warnings = this.get_bulk_edit_warnings(
-				state.headers,
-				state.rows,
-				state.row_numbers,
-				state.import_type,
-				map,
-			);
-			warnings.push(
-				...(await this.get_bulk_edit_link_warnings(state.rows, state.row_numbers, map)),
-			);
-			// a later mapping change already started its own refresh; let that one win
-			if (request_id !== preview_request_id) return;
-			state.warnings = warnings;
-			// no text summary — the red, editable cells are the only warning
-			// surface now; state.warnings still gates Apply below
-			sync_preview_errors(warnings);
-
-			set_footer();
-		};
-
-		const on_file = async (data) => {
-			if (cint(data.length) - BULK_EDIT_CSV_HEADER_ROWS > BULK_EDIT_MAX_ROWS) {
-				frappe.throw(
-					__("Cannot import table with more than {0} rows.", [BULK_EDIT_MAX_ROWS]),
-				);
-			}
-
-			state.headers = data[0] || [];
-			state.rows = [];
-			// kept alongside rows so a warning can name the line in the file
-			state.row_numbers = [];
-			data.slice(BULK_EDIT_CSV_HEADER_ROWS).forEach((row, i) => {
-				if (!row.some((v) => v)) return;
-				state.rows.push(row);
-				state.row_numbers.push(BULK_EDIT_CSV_HEADER_ROWS + i + 1);
-			});
-
-			if (!state.rows.length) {
-				frappe.msgprint({
-					message: __("There are no rows to import in this file."),
-					title: __("Nothing to Import"),
-					indicator: "orange",
-				});
-				return;
-			}
-
-			state.column_map = await this.get_bulk_edit_column_map(state.headers);
-			build_preview();
-			tabs.set_disabled(TAB_PREVIEW, false);
-			tabs.set_active(TAB_PREVIEW);
-		};
-
-		// ---------------------------------------------------------- footer
-
-		const download = () => {
-			const values = setup_form.get_values();
-			if (!values) return;
-			this.download_bulk_edit_template(
-				values.file_type,
-				values.fields,
-				values.export_records,
-			);
-		};
-
-		/**
-		 * The tab's forward action. Only the one that commits the rows is solid;
-		 * the steps on the way there stay quiet, so the weight in the footer
-		 * always points at the end of the flow.
-		 */
-		const set_action = (label, handler, { solid = false } = {}) => {
-			dialog.set_primary_action(label, handler);
-			if (!solid) {
-				frappe.ui.button.dress(dialog.get_primary_btn(), { label, variant: "subtle" });
-			}
-		};
-
-		/**
-		 * Every tab's actions are standard dialog actions, so they all land in
-		 * the footer's right group and never move as the tab changes.
-		 */
-		const set_footer = () => {
-			const active = tabs.get_active();
-			dialog.get_primary_btn().addClass("hide").prop("disabled", false);
-			const $secondary = dialog.get_secondary_btn().addClass("hide");
-
-			if (active === TAB_SETUP) {
-				dialog.set_secondary_action_label(__("Download Template"));
-				dialog.set_secondary_action(() => download());
-				$secondary.prop("disabled", !setup_form.get_value("fields")?.length);
-				if (can_import) {
-					set_action(__("Next"), () => {
-						tabs.set_disabled(TAB_UPLOAD, false);
-						tabs.set_active(TAB_UPLOAD);
-					});
-					// nothing on this tab should ever block moving on
-					dialog.get_primary_btn().prop("disabled", false);
-				}
-				return;
-			}
-
-			if (active === TAB_PREVIEW) {
-				// mapping and the red/editable cells are both right here —
-				// Apply is the only action this step needs
-				set_action(
-					__("Apply"),
-					() => {
-						dialog.hide();
-						const rows = state.rows.filter(
-							(_, r) => !state.skipped_rows.has(state.row_numbers[r]),
-						);
-						this.apply_bulk_edit_rows(rows, state.import_type, state.column_map);
-					},
-					{ solid: true },
-				);
-				// a skipped row isn't being imported, so what's wrong with it no
-				// longer stands in the way of the rest of the file — same rule the
-				// Data Import doctype applies (value_mapping.py: "Row warnings for
-				// user-skipped rows are ignored")
-				dialog.get_primary_btn().prop(
-					"disabled",
-					state.warnings.some((w) => w.blocking && !state.skipped_rows.has(cint(w.row))),
-				);
-				return;
-			}
-
-			// upload: a picked file needs uploading first; a parsed one (either
-			// source, via on_file) just moves on. Stays visible either way.
-			set_action(__("Next"), () => {
-				if (uploaded_file_count()) {
-					file_uploader.upload_files();
-					return;
-				}
-				tabs.set_active(TAB_PREVIEW);
-			});
-			dialog
-				.get_primary_btn()
-				.prop("disabled", !uploaded_file_count() && tab_defs[TAB_PREVIEW].disabled);
-		};
-
-		dialog.show();
-		// an import type is picked by default, so upload is reachable straight away
-		if (can_import) tabs.set_disabled(TAB_UPLOAD, false);
-		set_footer();
-	}
-
-	download_bulk_edit_template(file_type, fieldnames, export_records) {
-		const title = this.get_bulk_edit_title();
-		const data = this.get_bulk_edit_data(fieldnames, export_records);
-
-		if (file_type === "CSV") {
-			frappe.tools.downloadify(data, null, title);
-			return;
-		}
-
-		// the desk bundle cannot write xlsx, so the sheet is rendered on the server
-		open_url_post("/api/method/frappe.desk.form.bulk_edit.download_bulk_edit_template", {
-			doctype: this.frm.doctype,
-			title: title,
-			file_type: file_type,
-			data: JSON.stringify(data),
-		});
-	}
-
-	get_bulk_edit_data(fieldnames, export_records) {
-		let docfields = this.get_bulk_edit_docfields();
-		if (fieldnames && fieldnames.length) {
-			docfields = docfields.filter((df) => fieldnames.includes(df.fieldname));
-		}
-
-		// One header row of "Label (fieldname)" — a form get_bulk_edit_column_map
-		// resolves like any other, since it is one of the three the Data Import
-		// doctype's own matcher registers for every field. That doctype's template
-		// prints the fieldname only when two fields share a label (exporter.py
-		// add_header); here it is always printed, because a label alone can be
-		// read as the wrong column while filling the sheet, where there is no
-		// picker to say which field it feeds — "Item" and "Item Name" on Sales
-		// Invoice Item being the case that costs an import.
-		// The ID column is the exception: the matcher registers "ID" and "name" for
-		// it but never "ID (name)" (importer.py builds it from name_df, outside the
-		// loop that adds the parenthesised form), and a header it cannot resolve is
-		// a column that silently stops matching rows on Update.
-		const header = docfields.map((df) =>
-			df.fieldname === BULK_EDIT_ID_FIELDNAME
-				? __("ID")
-				: `${__(df.label || df.fieldname)} (${df.fieldname})`,
-		);
-		const data = [header];
-
-		let grid_rows = this.frm.doc[this.df.fieldname] || [];
-		if (export_records === BULK_EDIT_BLANK_TEMPLATE) grid_rows = [];
-		else if (export_records === BULK_EDIT_5_RECORDS) grid_rows = grid_rows.slice(0, 5);
-
-		grid_rows.forEach((d) => {
-			data.push(
-				docfields.map((df) => {
-					const value = d[df.fieldname];
-					if (!value) return "";
-					return df.fieldtype === "Date" ? frappe.datetime.str_to_user(value) : value;
-				}),
-			);
-		});
-
-		return data;
-	}
-
-	read_bulk_edit_file(file, on_parsed) {
-		// xlsx and xls need a reader the desk bundle does not have, and routing csv
-		// through the same call keeps every format producing identical rows
-		frappe.call({
-			method: "frappe.desk.form.bulk_edit.parse_bulk_edit_file",
-			args: {
-				doctype: this.frm.doctype,
-				filename: file.name,
-				dataurl: file.dataurl,
-			},
-			freeze: true,
-			freeze_message: __("Reading {0}", [file.name]),
-			callback: (r) => {
-				if (r.message) on_parsed(r.message);
-			},
-		});
-	}
-
-	// Same server-side fetch Data Import uses for a Google Sheets URL
-	// (frappe.utils.csvutils.get_csv_content_from_google_sheets), returning
-	// rows in the same shape read_bulk_edit_file does.
-	read_bulk_edit_google_sheet(url, on_parsed) {
-		frappe.call({
-			method: "frappe.desk.form.bulk_edit.parse_bulk_edit_google_sheet",
-			args: { doctype: this.frm.doctype, url },
-			freeze: true,
-			freeze_message: __("Reading Google Sheet"),
-			callback: (r) => {
-				if (r.message) on_parsed(r.message);
-			},
-		});
-	}
-
-	/**
-	 * The header keeps the file's own column name, so a column stays findable by
-	 * what the file calls it however it's mapped; the first body row is the
-	 * picker naming the field it lands in. Two different things, one above the
-	 * other, neither standing in for the other.
-	 */
-	get_bulk_edit_preview_html(headers, rows, row_numbers) {
-		const escape = frappe.utils.escape_html;
-		const shown = rows.slice(0, BULK_EDIT_PREVIEW_ROWS);
-
-		const head = headers.map(
-			(header, i) => `<th data-col="${i}" data-mapped="0">${escape(cstr(header))}</th>`,
-		);
-		// trailing column: the per-row Skip button, mounted by sync_preview_errors
-		// on the rows that need one. Header and mapping row carry an empty cell
-		// each so the columns stay aligned.
-		const mapping_row = `
-			<tr class="bulk-edit-mapping-row">
-				<td class="bulk-edit-preview-row"></td>
-				${headers.map((header, i) => `<td data-col="${i}"></td>`).join("")}
-				<td class="bulk-edit-skip-cell"></td>
-			</tr>
-		`;
-		const body = shown.map(
-			(row, r) => `
-				<tr data-row="${cint(row_numbers[r])}">
-					<td class="bulk-edit-preview-row">${cint(row_numbers[r])}</td>
-					${headers
-						.map(
-							(header, i) =>
-								`<td data-col="${i}" data-mapped="0">${escape(cstr(row[i]))}</td>`,
-						)
-						.join("")}
-					<td class="bulk-edit-skip-cell"></td>
-				</tr>
-			`,
-		);
-
-		return `
-			<div class="bulk-edit-preview-head">
-				<span class="text-muted small">${__(
-					"Map each column of the file to a field. Anything left unmapped is ignored.",
-				)}</span>
-				<span class="text-muted small">${
-					rows.length > shown.length
-						? __("Showing first {0} of {1} rows", [shown.length, rows.length])
-						: __("Showing all {0} rows", [rows.length])
-				}</span>
-			</div>
-			<div class="bulk-edit-preview-message text-muted small"></div>
-			<div class="bulk-edit-preview-table">
-				<table class="table table-bordered">
-					<thead>
-						<tr>
-							<th class="bulk-edit-preview-row">${__("Row")}</th>
-							${head.join("")}
-							<th class="bulk-edit-skip-cell"></th>
-						</tr>
-					</thead>
-					<tbody>${mapping_row}${body.join("")}</tbody>
-				</table>
-			</div>
-		`;
-	}
-
-	/**
-	 * Everything questionable about the file that can be found without touching the
-	 * server. Link existence is checked separately, in {@link get_bulk_edit_link_warnings}.
-	 */
-	get_bulk_edit_warnings(headers, rows, row_numbers, import_type, column_map) {
-		const warnings = [];
-		const columns = Object.keys(column_map).map(cint);
-		const id_index = columns.find((i) => column_map[i] === BULK_EDIT_ID_FIELDNAME);
-
-		headers.forEach((header, i) => {
-			if (header && column_map[i] === undefined) {
-				warnings.push({
-					col: i + 1,
-					message: __('"{0}" does not match a field and will be ignored.', [header]),
-				});
-			}
-		});
-
-		if (import_type !== BULK_EDIT_INSERT && id_index === undefined) {
-			warnings.push({
-				blocking: true,
-				message: __(
-					"No ID column is mapped, so no row can be matched. Every row will be skipped.",
-				),
-			});
-		} else if (id_index !== undefined) {
-			// rows sharing an ID target the same row; the last one applied wins
-			const rows_by_id = {};
-			rows.forEach((row, r) => {
-				const id = cstr(row[id_index]).trim();
-				if (id) (rows_by_id[id] ??= []).push(row_numbers[r]);
-			});
-			Object.entries(rows_by_id).forEach(([id, id_rows]) => {
-				if (id_rows.length > 1) {
-					warnings.push({
-						message: __("ID {0} appears in rows {1} — only the last one will apply.", [
-							id,
-							id_rows.join(", "),
-						]),
-					});
-				}
-			});
-		}
-
-		rows.forEach((row, r) => {
-			const row_number = row_numbers[r];
-
-			if (row.length !== headers.length) {
-				warnings.push({
-					row: row_number,
-					message:
-						row.length < headers.length
-							? __("This row has fewer cells than the header.")
-							: __("This row has more cells than the header."),
-				});
-			}
-
-			const id = id_index === undefined ? null : cstr(row[id_index]).trim();
-			const is_new =
-				import_type === BULK_EDIT_INSERT ||
-				(!this.get_bulk_edit_row_by_id(id) && import_type === BULK_EDIT_UPSERT);
-
-			if (import_type === BULK_EDIT_UPDATE && id && !this.get_bulk_edit_row_by_id(id)) {
-				warnings.push({
-					row: row_number,
-					message: __('No row in this table has the ID "{0}".', [id]),
-				});
-			}
-
-			columns.forEach((i) => {
-				const fieldname = column_map[i];
-				if (fieldname === BULK_EDIT_ID_FIELDNAME) return;
-
-				const df = frappe.meta.get_docfield(this.df.options, fieldname);
-				if (!df) return;
-
-				const value = cstr(row[i]).trim();
-
-				if (df.fieldtype === "Select" && value) {
-					const options = (df.options || "").split("\n").map((o) => o.trim());
-					if (!options.includes(value)) {
-						warnings.push({
-							blocking: true,
-							row: row_number,
-							col: i,
-							field: df,
-							message: __('"{0}" is not a valid option. Allowed: {1}', [
-								value,
-								options.filter(Boolean).join(", "),
-							]),
-						});
-					}
-				}
-
-				// Date/Datetime: same check the real Data Import doctype does
-				// (importer.py Row.validate_value), except that one guesses each
-				// column's own date format server-side from its values — no
-				// server round trip here, so this checks against the user's
-				// configured format and the system's internal one instead
-				if ((df.fieldtype === "Date" || df.fieldtype === "Datetime") && value) {
-					const user_fmt = frappe.datetime.get_user_date_fmt().toUpperCase();
-					const formats =
-						df.fieldtype === "Date"
-							? [user_fmt, frappe.defaultDateFormat]
-							: [
-									`${user_fmt} ${frappe.datetime.get_user_time_fmt()}`,
-									frappe.defaultDatetimeFormat,
-								];
-					if (!moment(value, formats, true).isValid()) {
-						warnings.push({
-							blocking: true,
-							row: row_number,
-							col: i,
-							field: df,
-							message:
-								df.fieldtype === "Date"
-									? __('"{0}" is not a valid date. Use {1}', [value, user_fmt])
-									: __('"{0}" is not a valid datetime. Use {1}', [
-											value,
-											`${user_fmt} ${frappe.datetime.get_user_time_fmt()}`,
-										]),
-						});
-					}
-				}
-
-				// exact port of importer.py's DURATION_PATTERN, plus the plain
-				// seconds the template itself exports
-				if (
-					df.fieldtype === "Duration" &&
-					value &&
-					!BULK_EDIT_DURATION_PATTERN.test(value) &&
-					!BULK_EDIT_SECONDS_PATTERN.test(value)
-				) {
-					warnings.push({
-						blocking: true,
-						row: row_number,
-						col: i,
-						field: df,
-						message: __('"{0}" is not valid. Use duration format: d h m s', [value]),
-					});
-				}
-
-				if (
-					df.fieldtype === "Time" &&
-					value &&
-					!moment(value, BULK_EDIT_TIME_FORMATS(), true).isValid()
-				) {
-					warnings.push({
-						blocking: true,
-						row: row_number,
-						col: i,
-						field: df,
-						message: __('"{0}" is not a valid time. Use {1}', [
-							value,
-							frappe.datetime.get_user_time_fmt(),
-						]),
-					});
-				}
-
-				// flt() and cint() turn anything unparseable into 0 without a word —
-				// the same blind spot importer.py Row.parse_value has — so "abc"
-				// would import as 0 and read afterwards as a real figure
-				if (
-					BULK_EDIT_NUMERIC_FIELDTYPES.includes(df.fieldtype) &&
-					value &&
-					!bulk_edit_is_number(value)
-				) {
-					warnings.push({
-						blocking: true,
-						row: row_number,
-						col: i,
-						field: df,
-						message: __('"{0}" is not a valid number.', [value]),
-					});
-				}
-
-				// a Check holds 0 or 1, and cint() reads every other word as 0 — so
-				// an unrecognised one would quietly import as unticked
-				if (
-					df.fieldtype === "Check" &&
-					value &&
-					!["0", "1", ...BULK_EDIT_CHECK_TRUE, ...BULK_EDIT_CHECK_FALSE].includes(
-						value.toLowerCase(),
-					)
-				) {
-					warnings.push({
-						blocking: true,
-						row: row_number,
-						col: i,
-						field: df,
-						message: __('"{0}" is not valid. Use {1}', [value, "0, 1, Yes, No"]),
-					});
-				}
-
-				// a blank mandatory cell only matters on a row that is being created
-				if (df.reqd && !value && is_new) {
-					warnings.push({
-						blocking: true,
-						row: row_number,
-						col: i,
-						field: df,
-						message: __("This field is mandatory and is blank."),
-					});
-				}
-			});
-		});
-
-		return warnings;
-	}
-
-	/** Existence-check every mapped Link column; batched into one call by target doctype. */
-	async get_bulk_edit_link_warnings(rows, row_numbers, column_map) {
-		const link_columns = Object.keys(column_map)
-			.map(cint)
-			.map((i) => ({ i, df: frappe.meta.get_docfield(this.df.options, column_map[i]) }))
-			.filter(({ df }) => df?.fieldtype === "Link");
-
-		if (!link_columns.length) return [];
-
-		const values_by_doctype = {};
-		link_columns.forEach(({ i, df }) => {
-			const values = (values_by_doctype[df.options] ??= new Set());
-			rows.forEach((row) => {
-				const value = cstr(row[i]).trim();
-				if (value) values.add(value);
-			});
-		});
-
-		const invalid = await frappe.xcall("frappe.desk.form.bulk_edit.get_invalid_link_values", {
-			doctype: this.frm.doctype,
-			values_by_doctype: JSON.stringify(
-				Object.fromEntries(
-					Object.entries(values_by_doctype).map(([doctype, values]) => [
-						doctype,
-						[...values],
-					]),
-				),
-			),
-		});
-
-		const warnings = [];
-		link_columns.forEach(({ i, df }) => {
-			const invalid_values = new Set(invalid[df.options] || []);
-			if (!invalid_values.size) return;
-			rows.forEach((row, r) => {
-				const value = cstr(row[i]).trim();
-				if (value && invalid_values.has(value)) {
-					warnings.push({
-						blocking: true,
-						row: row_numbers[r],
-						col: i,
-						field: df,
-						message: __('"{0}" is not a valid {1}', [value, df.label]),
-					});
-				}
-			});
-		});
-		return warnings;
-	}
-
-	/**
-	 * Column index to fieldname, for every header that names a field. Matched
-	 * server-side by the Data Import doctype's own header map, so a label, a
-	 * fieldname or "Label (fieldname)" all resolve — see bulk_edit.py.
-	 */
-	get_bulk_edit_column_map(headers) {
-		return frappe.xcall("frappe.desk.form.bulk_edit.get_bulk_edit_column_map", {
-			doctype: this.frm.doctype,
-			fieldname: this.df.fieldname,
-			headers: JSON.stringify(headers),
-		});
-	}
-
-	get_bulk_edit_row_by_id(id) {
-		return (this.frm.doc[this.df.fieldname] || []).find((d) => d.name === id);
-	}
-
-	apply_bulk_edit_rows(rows, import_type, column_map) {
-		const columns = Object.keys(column_map).map(cint);
-		const id_index = columns.find((i) => column_map[i] === BULK_EDIT_ID_FIELDNAME);
-		const counts = { insert: 0, update: 0, skip: 0 };
-
-		rows.forEach((row) => {
-			const id = id_index === undefined ? null : cstr(row[id_index]).trim();
-			const existing = id && this.get_bulk_edit_row_by_id(id);
-			let target;
-
-			if (import_type === BULK_EDIT_INSERT) {
-				target = this.frm.add_child(this.df.fieldname);
-				counts.insert++;
-			} else if (existing) {
-				target = existing;
-				counts.update++;
-			} else if (import_type === BULK_EDIT_UPSERT) {
-				target = this.frm.add_child(this.df.fieldname);
-				counts.insert++;
-			} else {
-				counts.skip++;
-				return;
-			}
-
-			columns.forEach((i) => {
-				const fieldname = column_map[i];
-				// the ID is what the row was matched on, never something to overwrite
-				if (fieldname === BULK_EDIT_ID_FIELDNAME) return;
-
-				const df = frappe.meta.get_docfield(this.df.options, fieldname);
-				if (!df) return;
-
-				const format = BULK_EDIT_VALUE_FORMATTERS[df.fieldtype];
-				target[fieldname] = format ? format(row[i]) : row[i];
-			});
-		});
-
-		this.frm.refresh_field(this.df.fieldname);
-		frappe.show_alert({
-			message: __("{0} added, {1} updated, {2} skipped — save to apply", [
-				counts.insert,
-				counts.update,
-				counts.skip,
-			]),
-			indicator: "green",
-		});
-
-		// add_child() and assigning to target[fieldname] both write straight to
-		// the local doc without going through frappe.model.set_value, so the form
-		// is never marked changed. Without this the rows would exist only in this
-		// browser and be lost on reload with no unsaved-changes warning.
-		this.frm.dirty();
 	}
 
 	add_custom_button(label, click, position = "bottom") {
